requirements.txt
streamlit
google-generativeai>=0.3.0
pdfplumber
scikit-learn
requests
aiohttp
faiss-cpu
numpy
pandas
sentence-transformers
python-dotenv
arxiv
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
pypdf
chromadb
mysql-connector-python>=8.0.26